pyproject.toml
[tool.poetry]
name = "events-tree-architect"
version = "0.1.0"
description = ""
authors = ["thetrueryan <georgytoper@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.3"
fastapi = {extras = ["standard"], version = "^0.116.1"}
pydantic-settings = "^2.10.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
pytest = "^8.4.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.3.0"
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"
black = "^25.1.0"
asyncpg = "^0.30.0"
python-dotenv = "^1.1.1"
pytest-asyncio = "^1.1.0"
docker = "^7.1.0"
pytest-dotenv = "^0.5.2"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
